backend/requirements.txt
appdirs==1.4.4
bcrypt==3.2.0
black==21.6b0
cffi==1.14.5
click==8.0.1
dnspython==2.1.0
Flask==2.0.1
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mypy-extensions==0.4.3
pathspec==0.8.1
pycparser==2.20
PyJWT==2.1.0
pymongo==3.11.4
python-dotenv==0.18.0
regex==2021.4.4
six==1.16.0
toml==0.10.2
Werkzeug==2.0.1
